+++ src/features/attachments/StoreAttachmentsInNode.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 
 import deepEqual from 'fast-deep-equal';
 
-import { useTaskOverrides } from 'src/core/contexts/taskStoreContext';
+import { useTaskOverrides } from 'src/core/contexts/TaskOverrides';
 import { useApplicationMetadata } from 'src/features/applicationMetadata/ApplicationMetadataProvider';
 import { isAttachmentUploaded } from 'src/features/attachments/index';
 import { DEFAULT_DEBOUNCE_TIMEOUT } from 'src/features/formData/types';
