+++ src/layout/Summary2/CommonSummaryComponents/EditButton.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { PencilIcon } from '@navikt/aksel-icons';
 
 import { Button } from 'src/app-components/Button/Button';
-import { useTaskOverrides } from 'src/core/contexts/taskStoreContext';
+import { useTaskOverrides } from 'src/core/contexts/TaskOverrides';
 import { useSetReturnToView, useSetSummaryNodeOfOrigin } from 'src/features/form/layout/PageNavigationContext';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
@@ -68,7 +68,7 @@ export function EditButton({
 
   const overrides = useTaskOverrides();
   const overriddenTaskId = overrides?.taskId;
-  const overriddenDataElementId = overrides?.dataElementId;
+  const overriddenDataElementId = overrides?.dataModelElementId;
   const indexedId = useIndexedId(targetBaseComponentId, skipLastIdMutator);
   const summary2Id = useSummaryProp('id');
