+++ src/layout/Summary2/SummaryComponent2/TaskSummaryWrapper.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TaskOverrides } from 'src/core/contexts/taskStoreContext';
+import { TaskOverrides } from 'src/core/contexts/TaskOverrides';
 import { FormProvider } from 'src/features/form/FormContext';
 import { useLayoutSets } from 'src/features/form/layoutSets/LayoutSetsProvider';
 import { useNavigationParam } from 'src/hooks/navigation';
